+++ src/models/index.ts
@@ -1,7 +1,0 @@
-export * from './Audio'
-export * from './Auth'
-export * from './Image'
-export * as IncidentModel from './Incident'
-export * from './Map'
-export * as ProjectModels from './Project'
-export * as StreamModels from './Stream'