+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.0.1 (2021-??-??)
 
+* **setup:** Fixed the player title label. Fixed the alignment of the navbar items ([CE-1480](https://jira.rfcx.org/browse/CE-1480))
+* **setup:** Updated the logic to return the dates difference into the dates util ([CE-1475](https://jira.rfcx.org/browse/CE-1475))
+* **setup:** Interfaces are added to all responses types ([CE-1476](https://jira.rfcx.org/browse/CE-1476))
 * **setup:** Display three events in the list of incidents([CE-1448](https://jira.rfcx.org/browse/CE-1448))
 * **setup:** The Ranger Track modal window is created ([CE-1446](https://jira.rfcx.org/browse/CE-1446))
 * **setup:** Implement UI for incidents and incident detail pages ([CE-1429](https://jira.rfcx.org/browse/CE-1429))